Generic RCS patch tweaks
* Change default fuel to Aerozine50/NTO
Rationale: Az50+NTO is used by most orbital maneuvering engines in stock
and has a thrust multiplier of 1, which simplifies the thrust
calculations for other fuels.
* Key the HAS block off of the absence of a ModuleEngineConfigs
Rationale: This will allow it to catch all unconfigured RCS thrusters.
NB: This ends up patching _all_ RCS thrusters to use hypergolics, which
may not be ideal for RCS thrusters that in stock will burn LFO. Separate
patches will be necessary for those thrusters.
* Change the mass multiplier to 0.8x to be consistent with Squad RCS